MCPサーバ起動
